게시판 서버 컴포넌트, 유저 컴포넌트 분리
